Move Api location for application to Student and change spac to can get doc from every student
