+++ src/application/application.service.ts
@@ -406,36 +406,4 @@ export class ApplicationService {
       semester: app.semester.semester,
     }));
   }
-
-  async GetApproveStudentDoc(studentId: string) {
-    try {
-      // Check if the student has an approved application
-      const application = await this.applicationRepository.findOne({
-        where: {
-          student: {
-            id: studentId,
-          },
-          adminApprovalTime: Not(IsNull()),
-        },
-        relations: {
-          student: true,
-        },
-      });
-
-      // If no approved application is found, return an error
-      if (!application) {
-        throw new Error(
-          `No approved application found for Student ID: ${studentId}`,
-        );
-      }
-
-      // Return the student's document details
-      return {
-        studentIdCardDocLink: application.student.studentIdCard,
-        studentbookBankDocLink: application.student.bookBank,
-      };
-    } catch (error) {
-      throw new Error(error.message || 'Error retrieving student document');
-    }
-  }
 }
