+++ src/student/student.service.ts
@@ -6,10 +6,11 @@ import {
 import { InjectRepository } from '@nestjs/typeorm';
 import { Student } from 'src/models/student.entity';
 import { S3Service } from 'src/s3/s3.service';
-import { Repository } from 'typeorm';
+import { IsNull, Not, Repository } from 'typeorm';
 import { UpdateStudentDto } from './dto/update-student.dto';
 import { UpdateStudentFilesDto } from './dto/update-student-files.dto';
 import { isNotEmptyObject } from 'class-validator';
+import { Application } from 'src/models/application.entity';
 
 @Injectable()
 export class StudentService {
@@ -17,6 +18,8 @@ export class StudentService {
     private readonly s3Service: S3Service,
     @InjectRepository(Student)
     private readonly studentRepository: Repository<Student>,
+    @InjectRepository(Application)
+    private readonly applicationRepository: Repository<Application>,
   ) {}
 
   async getStudent(userId: string) {
@@ -107,6 +110,40 @@ export class StudentService {
       if (!files.studentIdCard && !files.bookBank) {
         throw new UnprocessableEntityException('No data to update');
       }
+    }
+  }
+
+  async GetApproveStudentDoc(studentId: string) {
+    try {
+      // Check if the student has an approved application
+      const Astudent = await this.studentRepository.findOne({
+        where: {
+          id: studentId,
+        },
+      });
+
+      // If no student is found, return an NotFoundException
+      if (!Astudent) {
+        throw new NotFoundException(`No found for Student ID: ${studentId}`);
+      }
+
+      // Return the student's document details
+      return {
+        studentIDCardDocLink: Astudent.studentIdCard
+          ? await this.s3Service.getFileUrl(
+              'major-scholar-student-id-card',
+              Astudent.studentIdCard,
+            )
+          : null,
+        studentBookBankDocLink: Astudent.bookBank
+          ? await this.s3Service.getFileUrl(
+              'major-scholar-student-book-bank',
+              Astudent.bookBank,
+            )
+          : null,
+      };
+    } catch (error) {
+      throw new Error(error.message || 'Error retrieving student document');
     }
   }
 }
